+++ src/app/chat/chat.component.css
@@ -1,0 +1,5 @@
+#messages {
+    overflow: auto;
+    min-height: 400px;
+    max-height: 400px;
+}
