+++ src/app/chat/chat.component.html
@@ -1,0 +1,9 @@
+<div id="status"></div>
+<div id="chatbox">
+  <input type="text" id="username" placeholder="Enter name...">
+  <button id="clear" class="small-2 large-2 cell alert button">Clear</button>
+  <div class="card">
+    <div id="messages" class="card-section"></div>
+  </div>
+  <textarea id="textarea" placeholder="Enter message..."></textarea>
+</div>
