+++ src/app/initiative/initiative.component.html
@@ -1,0 +1,3 @@
+<p>
+  initiative works!
+</p>
