+++ src/app/initiative/initiative.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-initiative',
+  templateUrl: './initiative.component.html',
+  styleUrls: ['./initiative.component.css']
+})
+export class InitiativeComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
