+++ src/app/tabletop-center/tabletop-center.component.html
@@ -1,8 +1,8 @@
-<div class="grid-x">
-  <div class="cell large-3">
+<div id="tabletop-center">
+  <div id="center-utils">
     <app-utils></app-utils>
   </div>
-  <div class="cell large-9">
+  <div id="center-tabletop">
     <app-tabletop></app-tabletop>
   </div>
 </div>
