+++ src/app/utils/utils.component.css
@@ -1,15 +1,12 @@
-#messages {
-    overflow: auto;
-    min-height: 400px;
-    max-height: 400px;
-}
-
 #utils {
     margin: 0;
     height: 100%;
+    background:#1779ba; 
 }
 
-button {
+.button {
     margin: 0;
-    padding: 15px 5px 15px 5px;
+    padding: 15px 5px;
+    min-width: 50px;
+    width: 20%;
 }
